refactor(mars-theme): refactor processors and move theme to index
