--- packages/mars-theme/src/processors/index.js
@@ -1,3 +1,0 @@
-import image from "@frontity/html2react/processors/image";
-
-export default [image];
